Move and split database elements in into `event_store::database` (record, repository, query)
