+++ mithril-aggregator/src/event_store/database/mod.rs
@@ -3,5 +3,10 @@
 //! representation with their associated queries.
 
 pub mod migration;
+pub(crate) mod query;
+pub(crate) mod record;
+mod repository;
+
+pub use repository::*;
 #[cfg(test)]
 pub(crate) mod test_helper;
