+++ mithril-aggregator/src/event_store/event.rs
@@ -1,13 +1,6 @@
-use anyhow::anyhow;
 use chrono::{DateTime, Utc};
 
-use std::{collections::HashMap, sync::Arc};
-
-use mithril_common::StdResult;
-use mithril_persistence::sqlite::{
-    ConnectionExtensions, HydrationError, Projection, Query, SourceAlias, SqLiteEntity,
-    SqliteConnection, WhereCondition,
-};
+use std::collections::HashMap;
 
 /// Event that is sent from a thread to be persisted.
 #[derive(Debug, Clone)]
@@ -64,152 +57,4 @@ pub struct Event {
 
     /// the `content` of the original [EventMessage] this Event originates from.
     pub content: String,
-}
-
-impl SqLiteEntity for Event {
-    fn hydrate(row: sqlite::Row) -> Result<Self, HydrationError>
-    where
-        Self: Sized,
-    {
-        let created_at = &row.read::<&str, _>("created_at");
-
-        let myself = Self {
-            event_id: row.read::<i64, _>("event_id"),
-            created_at: DateTime::parse_from_rfc3339(created_at)
-                .map_err(|e| {
-                    HydrationError::InvalidData(format!(
-                        "Could not turn string '{created_at}' to rfc3339 Datetime. Error: {e}"
-                    ))
-                })?
-                .with_timezone(&Utc),
-            source: row.read::<&str, _>("source").to_string(),
-            action: row.read::<&str, _>("action").to_string(),
-            content: row.read::<&str, _>("content").to_string(),
-        };
-
-        Ok(myself)
-    }
-
-    fn get_projection() -> Projection {
-        let mut projection = Projection::default();
-        projection.add_field("event_id", "event_id", "int");
-        projection.add_field("created_at", "created_at", "string");
-        projection.add_field("source", "source", "string");
-        projection.add_field("action", "action", "string");
-        projection.add_field("content", "content", "string");
-
-        projection
-    }
-}
-
-struct InsertEventQuery {
-    condition: WhereCondition,
-}
-
-impl InsertEventQuery {
-    fn one(message: EventMessage) -> StdResult<Self> {
-        let condition = WhereCondition::new(
-            "(source, action, content, created_at) values (?*, ?*, ?*, ?*)",
-            vec![
-                sqlite::Value::String(message.source),
-                sqlite::Value::String(message.action),
-                sqlite::Value::String(format!(
-                    r#"{{"headers": {}, "content": {}}}"#,
-                    serde_json::to_string(&message.headers)?,
-                    message.content
-                )),
-                sqlite::Value::String(Utc::now().to_rfc3339()),
-            ],
-        );
-
-        Ok(Self { condition })
-    }
-}
-
-impl Query for InsertEventQuery {
-    type Entity = Event;
-
-    fn filters(&self) -> WhereCondition {
-        self.condition.clone()
-    }
-
-    fn get_definition(&self, data: &str) -> String {
-        let projection = Self::Entity::get_projection().expand(SourceAlias::default());
-
-        format!(r#"insert into event {data} returning {projection}"#)
-    }
-}
-
-/// The EventPersister is the adapter to persist EventMessage turning them into
-/// Event.
-pub struct EventPersister {
-    connection: Arc<SqliteConnection>,
-}
-
-impl EventPersister {
-    /// Instantiate an EventPersister
-    pub fn new(connection: Arc<SqliteConnection>) -> Self {
-        Self { connection }
-    }
-
-    /// Save an EventMessage in the database.
-    pub fn persist(&self, message: EventMessage) -> StdResult<Event> {
-        let log_message = message.clone();
-        let mut rows = self.connection.fetch(InsertEventQuery::one(message)?)?;
-
-        rows.next().ok_or(anyhow!(
-            "No record from the database after I saved event message {log_message:?}"
-        ))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::event_store::database::test_helper::event_store_db_connection;
-    use mithril_common::StdResult;
-
-    #[test]
-    fn event_projection() {
-        let projection = Event::get_projection();
-
-        assert_eq!(
-            "event_id as event_id, created_at as created_at, source as source, action as action, content as content".to_string(),
-            projection.expand(SourceAlias::default())
-        )
-    }
-
-    #[test]
-    fn provider_sql() {
-        let message = EventMessage::new("source", "action", "content");
-        let (parameters, values) = InsertEventQuery::one(message).unwrap().filters().expand();
-
-        assert_eq!(
-            "(source, action, content, created_at) values (?1, ?2, ?3, ?4)".to_string(),
-            parameters
-        );
-        assert_eq!(4, values.len());
-    }
-
-    #[test]
-    fn can_persist_event() -> StdResult<()> {
-        let connection = Arc::new(event_store_db_connection().unwrap());
-
-        let persister = EventPersister::new(connection);
-        let message = EventMessage::new("source", "action", "content");
-
-        let _event = persister.persist(message)?;
-        Ok(())
-    }
-
-    #[test]
-    fn migration_executed_running_database() -> StdResult<()> {
-        let connection = Arc::new(event_store_db_connection().unwrap());
-
-        let persister = EventPersister::new(connection);
-        let message = EventMessage::new("source", "action", "content");
-
-        let _event = persister.persist(message)?;
-        Ok(())
-    }
 }
